HealthTrackerApp/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
email-validator==2.1.0
matplotlib==3.8.0
numpy>=1.26.2
pandas>=2.2.0
pytest==7.4.0
gunicorn==21.2.0 